Can now fill the ellipse tool
Ellipses are hard, it turns out. I'm basically caching the min/max of
each row and filling in each tile (with a check to make sure I'm not
injecting extra, duplicate tiles, which would complicate the undo
feature pretty big when it comes time for that) if the tool is marked to
be filled.