Editor panel now looks for tiles below if no tile exists in current layer
Actually, if the current layer's tile is "default", since there is
ALWAYS a tile for every spot on every layer. But anyway. Also, it
filters the layers closer to white the further down they get using a
simple sine function based at 8 tiles for most filtered and a bit of a
boost up front.